Handle imports better and fix some bugs

+++ src/MSBuild.Abstractions/MSBuildConversionWorkspace.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.IO;
 using System.Linq;
+using System.Diagnostics.CodeAnalysis;
 
 using Microsoft.Build.Construction;
 using Microsoft.Build.Evaluation;
@@ -36,24 +37,27 @@ namespace MSBuild.Abstractions
                 var root = new MSBuildProjectRootElement(ProjectRootElement.Open(path, collection, preserveFormatting: true));
                 if (IsSupportedProjectType(root))
                 {
-                    if (!noBackup)
-                    {
-                        // Since git doesn't track the new '.old' addition in your changeset,
-                        // failing to overwrite will crash the tool if you have one in your directory.
-                        // This can be common if you're using the tool a few times and forget to delete the backup.
-                        File.Copy(path, path + ".old", overwrite: true);
-                    }
 
                     var configurations = DetermineConfigurations(root);
 
                     var unconfiguredProject = new UnconfiguredProject(configurations);
                     unconfiguredProject.LoadProjects(collection, globalProperties, path);
 
-                    var baseline = CreateSdkBaselineProject(path, unconfiguredProject.FirstConfiguredProject, root, configurations, tfm, keepCurrentTFMs);
-                    root.Reload(throwIfUnsavedChanges: false, preserveFormatting: true);
 
-                    var item = new MSBuildConversionWorkspaceItem(root, unconfiguredProject, baseline);
-                    items.Add(item);
+                    if (TryCreateSdkBaselineProject(path, unconfiguredProject.FirstConfiguredProject, root, configurations, tfm, keepCurrentTFMs, out var baseline))
+                    {
+                        if (!noBackup)
+                        {
+                            // Since git doesn't track the new '.old' addition in your changeset,
+                            // failing to overwrite will crash the tool if you have one in your directory.
+                            // This can be common if you're using the tool a few times and forget to delete the backup.
+                            File.Copy(path, path + ".old", overwrite: true);
+                        }
+
+                        root.Reload(throwIfUnsavedChanges: false, preserveFormatting: true);
+                        var item = new MSBuildConversionWorkspaceItem(root, unconfiguredProject, baseline.Value);
+                        items.Add(item);
+                    }
                 }
             }
 
@@ -79,7 +83,7 @@ namespace MSBuild.Abstractions
                                 builder.Add(dimensionValuePair.Value, dimensionValueDictionary.ToImmutableDictionary());
                             }
                         }
-                    }                   
+                    }
                 }
             }
             return builder.ToImmutable();
@@ -90,7 +94,7 @@ namespace MSBuild.Abstractions
         /// We need to use the same name as the original csproj and same path so that all the default that derive
         /// from name\path get the right values (there are a lot of them).
         /// </summary>
-        private BaselineProject CreateSdkBaselineProject(string projectFilePath, IProject project, IProjectRootElement root, ImmutableDictionary<string, ImmutableDictionary<string, string>> configurations, string tfm, bool keepCurrentTFMs)
+        private bool TryCreateSdkBaselineProject(string projectFilePath, IProject project, IProjectRootElement root, ImmutableDictionary<string, ImmutableDictionary<string, string>> configurations, string tfm, bool keepCurrentTFMs, [NotNullWhen(true)] out BaselineProject? baselineProject)
         {
             var projectStyle = GetProjectStyle(root);
             var outputType = GetProjectOutputType(root);
@@ -111,7 +115,8 @@ namespace MSBuild.Abstractions
                             : DesktopFacts.WinSDKAttribute; // pre-.NET 5 apps need a special SDK attribute.
                     break;
                 default:
-                    throw new NotSupportedException($"This project has custom imports in a manner that's not supported. '{projectFilePath}'");
+                    baselineProject = null;
+                    return false;
             }
 
             var propGroup = rootElement.AddPropertyGroup();
@@ -166,7 +171,8 @@ namespace MSBuild.Abstractions
                     ? MSBuildFacts.Net5Windows
                     : tfm;
 
-            return new BaselineProject(newProject, propertiesInTheBaseline, projectStyle, outputType, tfm, keepCurrentTFMs);
+            baselineProject = new BaselineProject(newProject, propertiesInTheBaseline, projectStyle, outputType, tfm, keepCurrentTFMs);
+            return true;
         }
 
         private bool IsSupportedOutputType(ProjectOutputType type) =>
@@ -231,21 +237,13 @@ namespace MSBuild.Abstractions
                 }
                 else
                 {
-                    var lastImport = imports.Last();
-                    var lastImportFileName = Path.GetFileName(lastImport.Project);
+                    var cleansedImports = imports.Select(import => Path.GetFileName(import.Project));
+                    var allImportsConvertibleToSdk =
+                        cleansedImports.All(import =>
+                            MSBuildFacts.PropsConvertibleToSDK.Contains(import, StringComparer.OrdinalIgnoreCase) ||
+                            MSBuildFacts.TargetsConvertibleToSDK.Contains(import, StringComparer.OrdinalIgnoreCase));
 
-                    if (firstImportFileName == FSharpFacts.FSharpTargetsPathVariableName)
-                    {
-                        firstImportFileName = Path.GetFileName(FSharpFacts.FSharpTargetsPath);
-                    }
-
-                    if (lastImportFileName == FSharpFacts.FSharpTargetsPathVariableName)
-                    {
-                        lastImportFileName = Path.GetFileName(FSharpFacts.FSharpTargetsPath);
-                    }
-
-                    if (MSBuildFacts.PropsConvertibleToSDK.Contains(firstImportFileName, StringComparer.OrdinalIgnoreCase) &&
-                        MSBuildFacts.TargetsConvertibleToSDK.Contains(lastImportFileName, StringComparer.OrdinalIgnoreCase))
+                    if (allImportsConvertibleToSdk)
                     {
                         if (MSBuildHelpers.IsNETFrameworkMSTestProject(projectRootElement))
                         {
@@ -262,6 +260,26 @@ namespace MSBuild.Abstractions
                     }
                     else
                     {
+                        Console.WriteLine("This project has custom imports that are not accepted by try-convert.");
+                        Console.WriteLine("Unexpected custom imports were found:");
+
+                        var customImports =
+                            cleansedImports.Where(import =>
+                                !(MSBuildFacts.PropsConvertibleToSDK.Contains(import, StringComparer.OrdinalIgnoreCase) ||
+                                  MSBuildFacts.TargetsConvertibleToSDK.Contains(import, StringComparer.OrdinalIgnoreCase)));
+
+                        foreach (var import in customImports)
+                        {
+                            Console.WriteLine($"\t{import}");
+                        }
+
+                        Console.WriteLine("The following imports are considered valid for conversion:");
+
+                        foreach (var import in MSBuildFacts.TargetsConvertibleToSDK.Union(MSBuildFacts.PropsConvertibleToSDK))
+                        {
+                            Console.WriteLine($"\t{import}");
+                        }
+
                         // It's something else, no idea what though
                         return ProjectStyle.Custom;
                     }
@@ -327,7 +345,7 @@ namespace MSBuild.Abstractions
                     {
                         return true;
                     }
-                }
+            }
 
             static void PrintGuidMessage(IEnumerable<string> allSupportedProjectTypeGuids, IEnumerable<string> allReadProjectTypeGuids)
             {
